misc/HtmlTestSuite/results/test-1338.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta content="width=device-width,initial-scale=1" name="viewport"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<link href="https://github.com/w3c/IntersectionObserver/issues/457" rel="help"/>
<style>
  div {
    width: 100px;
    height: 100px;
    background: blue;
    margin: 10px
  }
</style>
</head><body><div id="root"/>
<script>
let t = async_test("IntersectionObserver reports a (non-intersecting) entry if different-document from the doc");
let doc = document.implementation.createHTMLDocument("");
let target = doc.createElement("div");
doc.body.appendChild(target);
new IntersectionObserver(
  t.step_func_done(function(records) {
    assert_equals(records.length, 1);
    assert_false(records[0].isIntersecting);
  }),
  { root: document.querySelector("#root") }
).observe(target);
</script>
</body></html>
